Move stratify-github into Rails app
