--- stratify-github/Gemfile
@@ -1,4 +1,0 @@
-source "http://rubygems.org"
-
-# Specify your gem's dependencies in stratify-github.gemspec
-gemspec
