--- stratify-github/README.md
@@ -1,21 +1,0 @@
-# stratify-github
-
-## Dependencies
-
-Stratify is developed and tested with the following dependencies.
-
-* Ruby 1.9.3 (See `.ruby-version` for the specific patch level.)
-* MongoDB 2.0
-
-## Development
-
-To get set up for development on stratify-github, clone the repo, and ...
-
-    cd stratify/stratify-github
-    gem install bundler
-    bundle
-    rake
-
-## License
-
-Copyright 2012 Daemian Mack ([daemianmack.com](http://daemianmack.com)). Released under the MIT license. See the LICENSE file for further details.
