--- stratify-github/lib/stratify-github.rb
@@ -1,6 +1,0 @@
-require 'stratify-base'
-
-require 'stratify-github/activity'
-require 'stratify-github/collector'
-require 'stratify-github/engine'
-require 'stratify-github/version'
