--- stratify-github/lib/stratify-github/engine.rb
@@ -1,9 +1,0 @@
-require 'rails'
-
-module Stratify
-  module GitHub
-    class Engine < Rails::Engine
-      engine_name :stratify_github
-    end
-  end
-end
