--- stratify-github/lib/stratify-github/version.rb
@@ -1,5 +1,0 @@
-module Stratify
-  module GitHub
-    VERSION = "0.1.3"
-  end
-end
