--- stratify-github/spec/mongoid.yml
@@ -1,3 +1,0 @@
-persist_in_safe_mode: true
-host: localhost
-database: stratify_github_test
