+++ stratify-rails/lib/stratify/github.rb
@@ -1,0 +1,2 @@
+require 'stratify/github/activity'
+require 'stratify/github/collector'
