+++ stratify-rails/lib/stratify/github/assets/images/github-icon-16.png
