+++ stratify-rails/lib/stratify/github/assets/images/github-icon-24.png
