+++ stratify-rails/lib/stratify/github/collector.rb
