+++ stratify-rails/lib/stratify/github/events.rb
