+++ stratify-rails/lib/stratify/github/presenter.rb
