+++ stratify-rails/spec/lib/stratify/github/activity_spec.rb
