+++ stratify-rails/spec/lib/stratify/github/integration_spec.rb
