+++ stratify-rails/spec/lib/stratify/github/presenter_spec.rb
